# Remove data frame check in `df_equal_scalar()`