@DavisVaughan 2019-11-05
I actually think we should remove all of the checks in `df_equal_scalar()` (the length check and the names check as well). At the very least we are doing those checks way too many times (once per row).

We can get some _massive_ improvements from that, and I think it would be "safe" since for `vec_equal()` we have already `vec_cast_common()` `x` and `y` to a common data frame type. If we ever use `equal_scalar()` with data frames internally, we likely have already called `vec_cast_common()` on the inputs here as well (like in the dictionary function), and I think it is the responsibility of the caller to make sure the types are the same, not `df_equal_scalar()` (especially since we don't do these kinds of checks anywhere else).

These benefits would also carry into `pivot_longer()` and other tidyr friends.

Here are some quick benchmarks that are a result of removing these checks. `vec_match()` is particularly impressive.

https://github.com/r-lib/vctrs/blob/b4f7be4a47a4968caf31a4bae3e401651f1484d0/src/equal.c#L177-L191

```r
library(vctrs)

df <- data.frame(x = 1:1e6, y = 1:1e6)
```

```r
# before
bench::mark(
  vec_equal(df, df),
  iterations = 300
)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(df, df)   85.1ms   90.5ms      11.0    11.5MB     1.09

# after
bench::mark(
  vec_equal(df, df),
  iterations = 300
)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(df, df)   33.8ms   35.7ms      27.8    11.5MB     2.75
```

```r
# before
bench::mark(
  vec_unique_loc(df),
  iterations = 300
)
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique_loc(df)   84.2ms   97.2ms      10.2    23.6MB     25.0

# after
bench::mark(
  vec_unique_loc(df),
  iterations = 300
)
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique_loc(df)   59.8ms   65.2ms      15.3    23.6MB     37.4
```

```r
# before
bench::mark(
  vec_match(df, df), 
  iterations = 20
)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_match(df, df)    355ms    379ms      2.61    19.4MB     14.8

# after
bench::mark(
  vec_match(df, df), 
  iterations = 20
)
#> # A tibble: 1 x 6
#>   expression             min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>        <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_match(df, df)    217ms    230ms      4.39    19.4MB     24.9
```