@DavisVaughan 2019-11-05
Speed boost in `pivot_wider()` before and after this idea while using this tidyr PR https://github.com/tidyverse/tidyr/pull/790.

```r
library(tidyr)
library(dplyr, warn.conflicts = FALSE)
library(reshape2)

mydf <- expand_grid(
  case = sprintf("%03d", seq(1, 4000)),
  year = seq(1900, 2000),
  name = c("x", "y", "z")
) %>%
  mutate(value = rnorm(nrow(.)))

# before
bench::mark(
  pivot = pivot_wider(mydf, names_from = "name", values_from = "value"),
  spread = spread(mydf, name, value),
  dcast = dcast(mydf, case + year ~ name),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 pivot         587ms    624ms      1.58     117MB     3.45
#> 2 spread        431ms    497ms      2.04     431MB     9.72
#> 3 dcast         408ms    459ms      2.15     457MB     8.35

# after
bench::mark(
  pivot = pivot_wider(mydf, names_from = "name", values_from = "value"),
  spread = spread(mydf, name, value),
  dcast = dcast(mydf, case + year ~ name),
  iterations = 50
)
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 pivot         347ms    383ms      2.58     117MB     5.63
#> 2 spread        429ms    485ms      2.05     431MB     9.75
#> 3 dcast         413ms    459ms      2.15     457MB     8.33
```