@lionel- 2019-11-06
I agree it'd make sense to reach a point where we can assume a properly formed data frame in these low level routines.

But common type and casting is generic, so we can't make these assumptions without any check. Maybe `vec_cast()` should somehow perform a post-condition checks for data frames.