@DavisVaughan 2019-11-06
I'm not sure it makes sense to just special case data frames. Why is this example with a custom class built on top of a double any different? We can still make it fail by having a bad `vec_cast()` method, but I don't think it is vctrs's job to check for that.

``` r
library(vctrs)

x <- new_vctr(1, class = "bad_vctr")

# vec_ptype2 is as expected, common type is a bad_vctr
vec_ptype2.bad_vctr <- function(x, y, ...) {
  UseMethod("vec_ptype2.bad_vctr", y)
}

vec_ptype2.bad_vctr.double <- function(x, y, ...) {
  new_vctr(numeric(), class = "bad_vctr")
}

vec_ptype2.double.bad_vctr <- function(x, y, ...) {
  new_vctr(numeric(), class = "bad_vctr")
}

vec_cast.bad_vctr <- function(x, to, ...) {
  UseMethod("vec_cast.bad_vctr")
}

# but casting a double to a bad_vctr is specified incorrectly!
vec_cast.bad_vctr.double <- function(x, to, ...) {
  "hi"
}

vec_cast.bad_vctr.bad_vctr <- function(x, to, ...) {
  x
}

# makes sense
vec_ptype2(x, 1)
#> <bad_vctr[0]>
vec_ptype2(1, x)
#> <bad_vctr[0]>

# makes sense
vec_cast(x, x)
#> <bad_vctr[1]>
#> [1] 1

# WAT
vec_cast(1, x)
#> [1] "hi"

# meaning we get here...
vec_match(1, x)
#> Error in vec_match(1, x): STRING_PTR_RO() can only be applied to a 'character', not a 'double'

vec_match(x, 1)
#> [1] NA
```

<sup>Created on 2019-11-06 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>


I would really like it if we could assume that `vec_cast()` always fulfills the contract that `x` will end up with the type of `to`. If the user created method doesn't fulfill that contract, then the problem is on their side and it isn't the responsibility of vctrs to check for that. 

It just makes reasoning about the functions so much easier if we can 100% say "okay, I've casted `x` and `y` to a common type, so now I know I can make certain assumptions about them". Having to do these additional steps feels like a step away from that.