@hadley 2019-11-06
I agree with @DavisVaughan; it's not our responsibility if a user defined method doesn't fulfil the contract (although we should avoid crashing in C)