@lionel- 2019-11-06
@DavisVaughan Data frames have a propensity to pop up in bad state. We have had a lot of crash reports with dplyr because of this.

I agree that we should only have post conditions for things that make R crash, weird error messages are ok if the contract is not fulfilled.